refacto(front): split webhooks activity
