+++ front/lib/triggers/webhook.ts
@@ -1,24 +1,45 @@
 import type { IncomingHttpHeaders } from "node:http";
 
-import type { Authenticator } from "@app/lib/auth";
+import { toFileContentFragment } from "@app/lib/api/assistant/conversation/content_fragment";
+import { hasReachedPublicAPILimits } from "@app/lib/api/programmatic_usage_tracking";
+import { Authenticator } from "@app/lib/auth";
 import type { DustError } from "@app/lib/error";
 import { getWebhookRequestsBucket } from "@app/lib/file_storage";
+import { matchPayload, parseMatcherExpression } from "@app/lib/matcher";
 import { WebhookRequestModel } from "@app/lib/models/assistant/triggers/webhook_request";
 import type { WebhookRequestTriggerStatus } from "@app/lib/models/assistant/triggers/webhook_request_trigger";
 import { WebhookRequestTriggerModel } from "@app/lib/models/assistant/triggers/webhook_request_trigger";
 import { countActiveSeatsInWorkspaceCached } from "@app/lib/plans/usage/seats";
+import { TriggerResource } from "@app/lib/resources/trigger_resource";
+import { UserResource } from "@app/lib/resources/user_resource";
 import { WebhookRequestResource } from "@app/lib/resources/webhook_request_resource";
+import { WebhookSourcesViewResource } from "@app/lib/resources/webhook_sources_view_resource";
 import { FathomClient } from "@app/lib/triggers/built-in-webhooks/fathom/fathom_client";
+import { checkTriggerForExecutionPerDayLimit } from "@app/lib/triggers/common";
+import { launchAgentTriggerWorkflow } from "@app/lib/triggers/temporal/common/client";
 import { launchAgentTriggerWebhookWorkflow } from "@app/lib/triggers/temporal/webhook/client";
+import { concurrentExecutor } from "@app/lib/utils/async_utils";
 import {
   getTimeframeSecondsFromLiteral,
   rateLimiter,
 } from "@app/lib/utils/rate_limiter";
 import { verifySignature } from "@app/lib/webhookSource";
 import logger from "@app/logger/logger";
-import type { Result } from "@app/types";
-import { Err, normalizeError, Ok, removeNulls } from "@app/types";
-import type { TriggerType } from "@app/types/assistant/triggers";
+import { statsDClient } from "@app/logger/statsDClient";
+import type { ContentFragmentInputWithFileIdType, Result } from "@app/types";
+import {
+  assertNever,
+  Err,
+  isString,
+  normalizeError,
+  Ok,
+  removeNulls,
+} from "@app/types";
+import type {
+  TriggerType,
+  WebhookTriggerType,
+} from "@app/types/assistant/triggers";
+import { isWebhookTrigger } from "@app/types/assistant/triggers";
 import type {
   WebhookProvider,
   WebhookSourceForAdminType,
@@ -155,6 +176,466 @@ export async function checkWebhookRequestForRateLimit(
     }
     return new Ok(undefined);
   }
+
+  return new Ok(undefined);
+}
+
+export async function validateEventSubscription({
+  webhookSource,
+  headers,
+  body,
+  webhookRequest,
+  workspaceId,
+  webhookRequestId,
+}: {
+  webhookSource: {
+    provider: WebhookProvider | null;
+    subscribedEvents: string[];
+  };
+  headers: Record<string, string>;
+  body: Record<string, unknown>;
+  webhookRequest: WebhookRequestResource;
+  workspaceId: string;
+  webhookRequestId: number;
+}): Promise<
+  Result<
+    {
+      skipReason: string | null;
+      receivedEventValue: string | null;
+    },
+    Error
+  >
+> {
+  const { provider, subscribedEvents } = webhookSource;
+
+  if (!provider) {
+    return new Ok({ skipReason: null, receivedEventValue: null });
+  }
+
+  if (!WEBHOOK_PRESETS[provider]) {
+    const errorMessage = `No webhook preset found for provider: ${provider}.`;
+    await webhookRequest.markAsFailed(errorMessage);
+    logger.error(
+      {
+        workspaceId,
+        webhookRequestId,
+      },
+      errorMessage
+    );
+    return new Err(new Error(errorMessage));
+  }
+
+  const {
+    eventCheck,
+    event_blacklist: blacklist,
+    events,
+  } = WEBHOOK_PRESETS[provider];
+
+  if (!eventCheck) {
+    return new Ok({ skipReason: null, receivedEventValue: null });
+  }
+
+  const { type, field } = eventCheck;
+
+  let receivedEventValue: string | null = null;
+  switch (type) {
+    case "headers":
+      receivedEventValue = headers[field.toLowerCase()];
+      break;
+    case "body":
+      const bodyField = body[field];
+      receivedEventValue = isString(bodyField) ? bodyField : null;
+      break;
+    default:
+      assertNever(type);
+  }
+
+  if (!receivedEventValue) {
+    const errorMessage = `Unable to determine webhook event from ${type}.`;
+    await webhookRequest.markAsFailed(errorMessage);
+    logger.error(
+      {
+        workspaceId,
+        webhookRequestId,
+      },
+      errorMessage
+    );
+    return new Err(new Error(errorMessage));
+  }
+
+  if (blacklist && blacklist.includes(receivedEventValue)) {
+    await webhookRequest.markAsProcessed();
+    logger.info(
+      {
+        workspaceId,
+        webhookRequestId,
+        provider,
+        eventValue: receivedEventValue,
+      },
+      "Webhook event is blacklisted, ignoring."
+    );
+    return new Ok({ skipReason: "blacklisted", receivedEventValue });
+  }
+
+  if (
+    !events.map((e) => e.value).includes(receivedEventValue) ||
+    !subscribedEvents.includes(receivedEventValue)
+  ) {
+    const errorMessage =
+      "Webhook event not subscribed or not in preset. Potential cause: the events selection was manually modified on the service.";
+    await webhookRequest.markAsFailed(errorMessage);
+    logger.error(
+      {
+        workspaceId,
+        webhookRequestId,
+        eventValue: receivedEventValue,
+      },
+      errorMessage
+    );
+    return new Err(new Error(errorMessage));
+  }
+
+  return new Ok({ skipReason: null, receivedEventValue });
+}
+
+/**
+ * Checks rate limits for a webhook trigger, including workspace-level and trigger-specific limits.
+ * Returns null if no rate limit is hit, otherwise returns the error message.
+ */
+async function checkTriggerRateLimits({
+  auth,
+  trigger,
+  webhookRequest,
+  workspaceId,
+  webhookRequestId,
+  provider,
+}: {
+  auth: Authenticator;
+  trigger: WebhookTriggerType;
+  webhookRequest: WebhookRequestResource;
+  workspaceId: string;
+  webhookRequestId: number;
+  provider: WebhookProvider | null;
+}): Promise<string | null> {
+  /**
+   * Check for workspace-level rate limits
+   * - for fair use execution mode, check global rate limits
+   * - for programmatic usage mode, check public API limits
+   */
+  let workspaceRateLimitErrorMessage: string | null = null;
+  if (!trigger.executionMode || trigger.executionMode === "fair_use") {
+    const globalRateLimitRes = await checkWebhookRequestForRateLimit(auth);
+    if (globalRateLimitRes.isErr()) {
+      workspaceRateLimitErrorMessage = globalRateLimitRes.error.message;
+    }
+  } else {
+    const publicAPILimit = await hasReachedPublicAPILimits(auth, true);
+    if (publicAPILimit) {
+      workspaceRateLimitErrorMessage =
+        "Workspace has reached its public API limits for the current billing period.";
+    }
+  }
+
+  if (workspaceRateLimitErrorMessage !== null) {
+    await webhookRequest.markRelatedTrigger({
+      trigger,
+      status: "rate_limited",
+      errorMessage: workspaceRateLimitErrorMessage,
+    });
+
+    statsDClient.increment("webhook_workspace_rate_limit.hit.count", 1, [
+      `provider:${provider}`,
+      `workspace_id:${workspaceId}`,
+    ]);
+
+    logger.error(
+      { workspaceId, webhookRequestId },
+      workspaceRateLimitErrorMessage
+    );
+
+    return workspaceRateLimitErrorMessage;
+  }
+
+  const specificRateLimiterRes = await checkTriggerForExecutionPerDayLimit(
+    auth,
+    {
+      trigger,
+    }
+  );
+
+  if (specificRateLimiterRes.isErr()) {
+    const errorMessage = specificRateLimiterRes.error.message;
+    await webhookRequest.markRelatedTrigger({
+      trigger,
+      status: "rate_limited",
+      errorMessage,
+    });
+    logger.warn(
+      { workspaceId, webhookRequestId, triggerId: trigger.sId },
+      errorMessage
+    );
+
+    statsDClient.increment("webhook_trigger_rate_limit.hit.count", 1, [
+      `provider:${provider}`,
+      `workspace_id:${workspaceId}`,
+      `trigger_id:${trigger.sId}`,
+    ]);
+
+    return errorMessage;
+  }
+
+  return null;
+}
+
+/**
+ * Checks if the webhook payload matches the trigger's filter expression.
+ * Returns true if it matches or if there is no filter.
+ */
+function applyPayloadFilter({
+  trigger,
+  body,
+  provider,
+  workspaceId,
+}: {
+  trigger: WebhookTriggerType;
+  body: Record<string, unknown>;
+  provider: WebhookProvider | null;
+  workspaceId: string;
+}): boolean {
+  const { filter } = trigger.configuration;
+
+  if (!filter) {
+    return true;
+  }
+
+  const tags = [
+    `provider:${provider}`,
+    `workspace_id:${workspaceId}`,
+    `trigger_id:${trigger.sId}`,
+  ];
+  statsDClient.increment("webhook_filter.events_processed.count", 1, tags);
+
+  const parsedFilterResult = parseMatcherExpression(filter);
+  if (parsedFilterResult.isErr()) {
+    logger.error(
+      {
+        triggerId: trigger.id,
+        triggerName: trigger.name,
+        filter,
+        err: parsedFilterResult.error,
+      },
+      "Invalid filter expression in webhook trigger"
+    );
+    return false;
+  }
+
+  const payloadMatchesFilter = matchPayload(body, parsedFilterResult.value);
+  if (payloadMatchesFilter) {
+    statsDClient.increment("webhook_filter.events_passed.count", 1, tags);
+    return true;
+  }
+
+  return false;
+}
+
+export async function filterTriggers({
+  auth,
+  webhookSource,
+  receivedEventValue,
+  webhookRequest,
+  body,
+}: {
+  auth: Authenticator;
+  webhookSource: { id: number; provider: WebhookProvider | null };
+  receivedEventValue: string | null;
+  webhookRequest: WebhookRequestResource;
+  body: Record<string, unknown>;
+}): Promise<Result<WebhookTriggerType[], Error>> {
+  const workspaceId = auth.getNonNullableWorkspace().sId;
+  const webhookRequestId = webhookRequest.id;
+  const { provider } = webhookSource;
+  // Fetch all triggers based on the webhook source id.
+  const views = await WebhookSourcesViewResource.listByWebhookSource(
+    auth,
+    webhookSource.id
+  );
+
+  // Fetch all triggers based on the webhook source id and flatten the result.
+  const triggers = (
+    await concurrentExecutor(
+      views,
+      async (view) => {
+        return TriggerResource.listByWebhookSourceViewId(auth, view.id);
+      },
+      { concurrency: 10 }
+    )
+  )
+    .flat()
+    .map((t) => t.toJSON())
+    // Filter here to avoid a lot of type checking later.
+    .filter(isWebhookTrigger)
+    // Filter out disabled triggers
+    .filter((t) => t.enabled);
+
+  const filteredTriggers: WebhookTriggerType[] = [];
+
+  for (const trigger of triggers) {
+    const { event } = trigger.configuration;
+
+    if (event && receivedEventValue && event !== receivedEventValue) {
+      // Received event doesn't match the trigger's event, skip this trigger
+      await webhookRequest.markRelatedTrigger({
+        trigger,
+        status: "not_matched",
+      });
+      continue;
+    }
+
+    const rateLimitError = await checkTriggerRateLimits({
+      auth,
+      trigger,
+      webhookRequest,
+      workspaceId,
+      webhookRequestId,
+      provider,
+    });
+
+    if (rateLimitError !== null) {
+      // If it's a workspace-level rate limit, return error immediately
+      if (rateLimitError.includes("Workspace")) {
+        return new Err(new Error(rateLimitError));
+      }
+      // Otherwise it's a trigger-specific rate limit, continue to next trigger
+      continue;
+    }
+
+    const matchesFilter = applyPayloadFilter({
+      trigger,
+      body,
+      provider,
+      workspaceId,
+    });
+
+    if (matchesFilter) {
+      filteredTriggers.push(trigger);
+    }
+  }
+
+  return new Ok(filteredTriggers);
+}
+
+export async function launchTriggersWorkflows({
+  auth,
+  filteredTriggers,
+  webhookSource,
+  body,
+  webhookRequest,
+}: {
+  auth: Authenticator;
+  filteredTriggers: WebhookTriggerType[];
+  webhookSource: { id: number };
+  body: Record<string, unknown>;
+  webhookRequest: WebhookRequestResource;
+}): Promise<Result<void, Error>> {
+  const workspaceId = auth.getNonNullableWorkspace().sId;
+  const webhookRequestId = webhookRequest.id;
+  // Check if any of the triggers requires the payload.
+  const requiresPayload = filteredTriggers.some(
+    (t) => t.configuration.includePayload
+  );
+
+  // If we need the payload, create a content fragment for it.
+  let contentFragment: ContentFragmentInputWithFileIdType | undefined;
+  if (requiresPayload) {
+    const contentFragmentRes = await toFileContentFragment(auth, {
+      contentFragment: {
+        contentType: "application/json",
+        content: JSON.stringify(body),
+        title: `Webhook body (source id: ${webhookSource.id}, date: ${new Date().toISOString()})`,
+      },
+      fileName: `webhook_body_${webhookSource.id}_${Date.now()}.json`,
+    });
+
+    if (contentFragmentRes.isErr()) {
+      const errorMessage =
+        "Error creating file content fragment from webhook request.";
+      await webhookRequest.markAsFailed(errorMessage);
+      logger.error({ workspaceId, webhookRequestId }, errorMessage);
+      return new Err(new Error(errorMessage));
+    }
+
+    contentFragment = contentFragmentRes.value;
+  }
+
+  // Launch all the triggers' workflows concurrently.
+  await concurrentExecutor(
+    filteredTriggers,
+    async (trigger) => {
+      // Get the trigger's user and create a new authenticator
+      const user = await UserResource.fetchByModelId(trigger.editor);
+
+      if (!user) {
+        logger.error(
+          {
+            triggerId: trigger.sId,
+          },
+          "Trigger editor not found."
+        );
+        await webhookRequest.markRelatedTrigger({
+          trigger,
+          status: "workflow_start_failed",
+        });
+      } else {
+        const auth = await Authenticator.fromUserIdAndWorkspaceId(
+          user.sId,
+          workspaceId
+        );
+        if (trigger.configuration.includePayload && !contentFragment) {
+          const errorMessage =
+            "One of the triggers requires the payload, but the contentFragment is missing. It should never happen as the content fragment is created if any of the triggers requires the payload.";
+          logger.error(
+            {
+              triggerId: trigger.sId,
+            },
+            errorMessage
+          );
+          await webhookRequest.markRelatedTrigger({
+            trigger,
+            status: "workflow_start_failed",
+          });
+          return;
+        }
+
+        // Fire and forget
+        const result = await launchAgentTriggerWorkflow({
+          auth,
+          trigger,
+          contentFragment,
+          webhookRequestId,
+        });
+
+        if (result.isErr()) {
+          await webhookRequest.markRelatedTrigger({
+            trigger,
+            status: "workflow_start_failed",
+          });
+          logger.error(
+            {
+              triggerId: trigger.sId,
+              error: result.error,
+            },
+            "Error launching agent trigger workflow."
+          );
+        } else {
+          await webhookRequest.markRelatedTrigger({
+            trigger,
+            status: "workflow_start_succeeded",
+          });
+        }
+      }
+    },
+    { concurrency: 10 }
+  );
 
   return new Ok(undefined);
 }
